Remove unnecessary f's from stdlib cmath functions
powf() is a C function but std::pow() is the thin C++ wrapper that
delegates to pow(), powf(), or powl() dependent on argument type.
Most of these are already acting on floats, and those that aren't have
been fixed up to do so.

+++ src/engine_view.cpp
@@ -40,7 +40,7 @@ void EngineView::onDrag(const Point &p0, const Point &mouse0, const Point &mouse
 }
 
 void EngineView::onMouseScroll(int scroll) {
-    const float f = std::powf(2.0, (float)scroll / 500.0f);
+    const float f = std::pow(2.0f, (float)scroll / 500.0f);
 
     const Point prevCenter = getCenter();
